+++ tests/Method/SetChatMemberTagTest.php
@@ -1,0 +1,56 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phptg\BotApi\Tests\Method;
+
+use PHPUnit\Framework\TestCase;
+use Phptg\BotApi\Method\SetChatMemberTag;
+use Phptg\BotApi\Transport\HttpMethod;
+use Phptg\BotApi\Tests\Support\TestHelper;
+
+use function PHPUnit\Framework\assertSame;
+use function PHPUnit\Framework\assertTrue;
+
+final class SetChatMemberTagTest extends TestCase
+{
+    public function testBase(): void
+    {
+        $method = new SetChatMemberTag(1, 2);
+
+        assertSame(HttpMethod::POST, $method->getHttpMethod());
+        assertSame('setChatMemberTag', $method->getApiMethod());
+        assertSame(
+            [
+                'chat_id' => 1,
+                'user_id' => 2,
+            ],
+            $method->getData(),
+        );
+    }
+
+    public function testFull(): void
+    {
+        $method = new SetChatMemberTag(1, 2, 'test');
+
+        assertSame(HttpMethod::POST, $method->getHttpMethod());
+        assertSame('setChatMemberTag', $method->getApiMethod());
+        assertSame(
+            [
+                'chat_id' => 1,
+                'user_id' => 2,
+                'tag' => 'test',
+            ],
+            $method->getData(),
+        );
+    }
+
+    public function testPrepareResult(): void
+    {
+        $method = new SetChatMemberTag(1, 2);
+
+        $preparedResult = TestHelper::createSuccessStubApi(true)->call($method);
+
+        assertTrue($preparedResult);
+    }
+}